requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
deprecation==2.1.0
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
frozenlist==1.5.0
gotrue==2.9.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
multidict==6.1.0
numpy==2.1.2
packaging==24.1
pillow==11.0.0
postgrest==0.17.2
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.0.6
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
storage3==0.8.2
supabase==2.9.1
supafunc==0.6.2
typing_extensions==4.12.2
websockets==13.1
yarl==1.17.1
